+++ src/main/groovy/org/jetbrains/intellij/tasks/PublishTask.groovy
@@ -9,8 +9,7 @@ import org.gradle.api.internal.ConventionTask
 import org.gradle.api.tasks.*
 import org.gradle.util.CollectionUtils
 import org.jetbrains.intellij.Utils
-import org.jetbrains.intellij.pluginRepository.PluginRepositoryInstance
-import retrofit.RetrofitError
+import org.jetbrains.intellij.pluginRepository.PluginRepositoryFactory
 
 class PublishTask extends ConventionTask {
     private Object distributionFile
@@ -126,18 +125,11 @@ class PublishTask extends ConventionTask {
             for (String channel : channels) {
                 Utils.info(this, "Uploading plugin ${pluginId} from $distributionFile.absolutePath to $host, channel: $channel")
                 try {
-                    def repoClient = new PluginRepositoryInstance(host, getToken())
-                    repoClient.uploadPlugin(pluginId, distributionFile, channel && 'default' != channel ? channel : '')
+                    def repoClient = PluginRepositoryFactory.create(host, getToken())
+                    repoClient.uploader.uploadPlugin(pluginId, distributionFile, channel && 'default' != channel ? channel : null, null)
                     Utils.info(this, "Uploaded successfully")
                 }
                 catch (exception) {
-                    def cause = exception.getCause()
-                    if (cause instanceof RetrofitError && cause.response && cause.response.status == 404) {
-                        def message = '''Failed to upload plugin
-The plugin should be uploaded manually at least once before using `publishPlugin` task.
-Please follow the instructions: https://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/publishing_plugin.html'''
-                        throw new TaskExecutionException(this, new GradleException(message))
-                    }
                     throw new TaskExecutionException(this, new GradleException('Failed to upload plugin', exception))
                 }
             }