+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Fix platform prefix for DataGrip [#458](../../issues/458)
 - Enable plugin auto-reloading by default
+- Upgrade plugins repository client
 
 ## 0.4.16